[es6-bookreader-core/DebugConsole]  rename debug console
