+++ src/js/BookReader/DebugConsole.js
